Update to use CEu buildscript
