+++ gradle.properties.old
@@ -1,0 +1,54 @@
+# Mod Information
+archives_base_name = multiblocktweaker
+
+mc_version = 1.12.2
+maven_group = eutros
+phase=stable-gtceu
+
+major=1
+minor=8
+patch=2
+
+
+## Hard Dependencies
+ccl_pid = 242818
+ccl_fid = 2779848
+gt_pid = 557242
+gt_fid = 4527757
+gt_sources_fid = 4527758
+
+## Soft Dependencies
+jei_version = 4.16.1.302
+crt_version = 4.1.20.684
+### TOP 1.4.28
+top_pid = 245211
+top_fid = 2667280
+### CTM 1.0.2.31
+ctm_pid = 267602
+ctm_fid = 2915363
+### AE2 0.55.6
+ae2_pid = 570458
+ae2_fid = 4402048
+
+## Mixin Dependencies
+mixinbooter_version = 7.0
+mixin_version = 0.8.3
+asm_debug_version = 5.2
+### should use 24.1.1 but 30.0+ has a vulnerability fix
+guava_version = 30.0
+### should use 2.8.6 but 2.8.9+ has a vulnerability fix
+gson_version = 2.8.9
+
+## Gradle Dependencies
+foojay_version = 0.4.0
+idea_ext_version = 1.1.7
+rfg_version = 1.3.+
+java_version = 8
+
+# Gradle Settings (do not rename values)
+org.gradle.welcome = never
+org.gradle.logging.stacktrace = all
+## Sets default memory used for gradle commands. Can be overridden by user or command line properties.
+## This is required to provide enough memory for the Minecraft decompilation process.
+org.gradle.jvmargs = -Xmx3G
+org.gradle.daemon = false
