+++ src/main/java/eutros/multiblocktweaker/MultiblockTweaker.java
@@ -1,13 +1,13 @@
 package eutros.multiblocktweaker;
 
-import eutros.MBTInternalTags;
+import com.myname.mymodid.MBTTags;
 import gregtech.GTInternalTags;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 
 @Mod(modid = MultiblockTweaker.MOD_ID,
         name = "Multiblock Tweaker",
-        version = MBTInternalTags.VERSION,
+        version = MBTTags.VERSION,
         dependencies = GTInternalTags.DEP_VERSION_STRING)
 public class MultiblockTweaker {
